+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/enums/add-comments-step.enum.ts
@@ -1,4 +1,5 @@
 export enum AddCommentsStep {
   HINT_TEXT = 'Explain why you are creating this flag. Do not include any sensitive information such as personal details.',
-  CHARACTER_LIMIT_INFO = 'You can enter up to 200 characters'
+  CHARACTER_LIMIT_INFO = 'You can enter up to 200 characters',
+  WARNING_TEXT = 'The details entered here MAY be visible to the party in the future.'
 }